Add content key back to attachments
- this is a result of some odd behavior from the bill.com api
- yesterday when adding a document without the `document` key, I got an
`invalid document` error, which was corrected by adding the document
key.  Confirmed that the document then showed up in the bill.com
sandbox.  Today if I upload with the document key, I do not receive an
error, but I end up with an empty document.  As it stands today we drop
the document key, but I'm hesitant to do so given that only yesterday
the api was seemingly changed without warning to require the `document`
key in lieu of `content`.